Update glue script to accommodate SuperGLUE evals
