+++ sequence_classification.py
@@ -10,7 +10,7 @@ from typing import Optional, cast
 # Add folder root to path to allow us to use relative imports regardless of what directory the script is run from
 sys.path.append(os.path.dirname(os.path.realpath(__file__)))
 
-import src.glue.data as data_module
+import src.evals.data as data_module
 import src.hf_bert as hf_bert_module
 import src.mosaic_bert as mosaic_bert_module
 import transformers
