+++ src/evals/finetuning_jobs.py
@@ -1,0 +1,245 @@
+# Copyright 2022 MosaicML Examples authors
+# SPDX-License-Identifier: Apache-2.0
+
+# """Contains GLUE job objects for the simple_glue_trainer."""
+import atexit
+import copy
+import gc
+import multiprocessing as mp
+import os
+import sys
+from multiprocessing import managers
+from typing import Any, Dict, List, Optional, Union, cast
+
+# Add glue folder root to path to allow us to use relative imports regardless of what directory the script is run from
+sys.path.append(os.path.dirname(os.path.realpath(__file__)))
+
+import torch
+from composer import ComposerModel
+from composer.core import Callback
+from composer.core.types import Dataset
+from composer.devices import Device, DeviceGPU
+from composer.loggers import LoggerDestination
+from composer.optim import ComposerScheduler
+from composer.trainer.trainer import Trainer
+from composer.utils import dist, reproducibility
+from torch.utils.data import DataLoader
+
+
+def _build_dataloader(dataset, **kwargs):
+    import transformers
+
+    dataset = cast(Dataset, dataset)
+
+    return DataLoader(
+        dataset=dataset,
+        sampler=dist.get_sampler(dataset, drop_last=False, shuffle=False),
+        collate_fn=transformers.default_data_collator,
+        **kwargs,
+    )
+
+
+Metrics = Dict[str, Dict[str, Any]]
+
+
+def reset_trainer(trainer: Trainer, garbage_collect: bool = False):
+    """Cleans up memory usage left by trainer."""
+    trainer.close()
+    # Unregister engine from atexit to remove ref
+    atexit.unregister(trainer.engine._close)
+    # Close potentially persistent dataloader workers
+    loader = trainer.state.train_dataloader
+    if loader and loader._iterator is not None:  # type: ignore
+        loader._iterator._shutdown_workers()  # type: ignore
+    # Explicitly delete attributes of state as otherwise gc.collect() doesn't free memory
+    for key in list(trainer.state.__dict__.keys()):
+        delattr(trainer.state, key)
+    # Delete the rest of trainer attributes
+    for key in list(trainer.__dict__.keys()):
+        delattr(trainer, key)
+    if garbage_collect:
+        gc.collect()
+        torch.cuda.empty_cache()
+
+
+class FineTuneJob:
+    """Encapsulates a fine-tuning job.
+
+    Tasks should subclass FineTuneJob and implement the
+    get_trainer() method.
+
+    Args:
+        name (str, optional): job name. Defaults to the class name.
+        load_path (str, optional): path to load checkpoints. Default: None
+        save_folder (str, optional): path to save checkpoints. Default: None
+        kwargs (dict, optional): additional arguments passed available to the Trainer.
+    """
+
+    def __init__(
+        self,
+        job_name: Optional[str] = None,
+        load_path: Optional[str] = None,
+        save_folder: Optional[str] = None,
+        seed: int = 42,
+        **kwargs,
+    ):
+        reproducibility.seed_all(seed)
+        self._job_name = job_name
+        self.seed = seed
+        self.load_path = load_path
+        self.save_folder = save_folder
+        self.kwargs = kwargs
+
+    def get_trainer(self, device: Optional[Union[str, Device]]) -> Trainer:
+        """Returns the trainer for the job."""
+        raise NotImplementedError
+
+    def print_metrics(self, metrics: Metrics):
+        """Prints fine-tuning results."""
+        job_name = self.job_name
+
+        print(f"Results for {job_name}:")
+        print("-" * (12 + len(job_name)))
+        for eval, metric in metrics.items():
+            for metric_name, value in metric.items():
+                print(f"{eval}: {metric_name}, {value*100:.2f}")
+        print("-" * (12 + len(job_name)))
+
+    @property
+    def job_name(self) -> str:
+        """Job name, defaults to class name."""
+        if self._job_name is not None:
+            return self._job_name
+        return self.__class__.__name__
+
+    def run(
+        self,
+        gpu_queue: Optional[mp.Queue] = None,
+        process_to_gpu: Optional[managers.DictProxy] = None,
+    ) -> Dict[str, Any]:
+        """Trains the model, optionally pulling a GPU id from the queue.
+
+        Returns:
+            A dict with keys:
+            * 'checkpoints': list of saved_checkpoints, if any,
+            * 'metrics': nested dict of results, accessed by
+                        dataset and metric name, e.g.
+                        ``metrics['glue_mnli']['MulticlassAccuracy']``.
+        """
+        if gpu_queue is None:
+            if torch.cuda.device_count() > 0:
+                gpu_id = 0
+                device = DeviceGPU(gpu_id)
+            else:
+                gpu_id = None
+                device = "cpu"
+        else:
+            current_pid = os.getpid()
+            assert process_to_gpu is not None
+            if current_pid in process_to_gpu:
+                gpu_id = process_to_gpu[current_pid]
+            else:
+                gpu_id = gpu_queue.get()
+                process_to_gpu[current_pid] = gpu_id
+            device = DeviceGPU(gpu_id)
+
+        print(f"Running {self.job_name} on GPU {gpu_id}")
+
+        trainer = self.get_trainer(device=device)
+
+        trainer.fit()
+
+        collected_metrics: Dict[str, Dict[str, Any]] = {}
+        for eval_name, metrics in trainer.state.eval_metrics.items():
+            collected_metrics[eval_name] = {
+                name: metric.compute().cpu().numpy() for name, metric in metrics.items()
+            }
+
+        saved_checkpoints = copy.copy(trainer.saved_checkpoints)
+        reset_trainer(trainer, garbage_collect=True)
+
+        self.print_metrics(collected_metrics)
+
+        output = {
+            "checkpoints": saved_checkpoints,
+            "metrics": collected_metrics,
+            "job_name": self.job_name,
+        }
+
+        return output
+
+
+class ClassificationJob(FineTuneJob):
+    def __init__(
+        self,
+        model: ComposerModel,
+        tokenizer_name: str,
+        job_name: Optional[str] = None,
+        seed: int = 42,
+        task_name: Optional[str] = None,
+        eval_interval: str = "1000ba",
+        scheduler: Optional[ComposerScheduler] = None,
+        max_sequence_length: Optional[int] = 256,
+        max_duration: Optional[str] = "3ep",
+        batch_size: Optional[int] = 32,
+        load_path: Optional[str] = None,
+        save_folder: Optional[str] = None,
+        loggers: Optional[List[LoggerDestination]] = None,
+        callbacks: Optional[List[Callback]] = None,
+        precision: Optional[str] = None,
+        **kwargs,
+    ):
+        if task_name is None:
+            raise ValueError(
+                "GlueClassificationJob should not be instantiated directly. Please instantiate a specific glue job type instead (e.g. MNLIJob)."
+            )
+        super().__init__(job_name, load_path, save_folder, seed, **kwargs)
+
+        self.task_name = task_name
+
+        self.eval_interval = eval_interval
+        self.tokenizer_name = tokenizer_name
+        self.model = model
+
+        self.scheduler = scheduler
+
+        self.max_sequence_length = max_sequence_length
+        self.max_duration = max_duration
+        self.batch_size = batch_size
+        self.loggers = loggers
+        self.callbacks = callbacks
+        self.precision = precision
+
+        # These will be set by the subclasses for specific GLUE tasks
+        self.train_dataloader = None
+        self.evaluators = None
+        self.optimizer = None
+
+    def get_trainer(self, device: Optional[Union[Device, str]] = None):
+        return Trainer(
+            model=self.model,
+            optimizers=self.optimizer,
+            schedulers=self.scheduler,
+            train_dataloader=self.train_dataloader,
+            eval_dataloader=self.evaluators,
+            eval_interval=self.eval_interval,
+            load_path=self.load_path,
+            save_folder=self.save_folder,
+            max_duration=self.max_duration,
+            seed=self.seed,
+            device_train_microbatch_size=(
+                "auto" if torch.cuda.device_count() > 0 else None
+            ),
+            load_weights_only=True,
+            load_strict_model_weights=False,
+            loggers=self.loggers,
+            callbacks=self.callbacks,
+            python_log_level="ERROR",
+            run_name=self.job_name,
+            load_ignore_keys=["state/model/model.classifier*"],
+            precision=self.precision,
+            device=device,
+            progress_bar=True,
+            log_to_console=False,
+            **self.kwargs,
+        )
