--- src/superglue/__init__.py
@@ -1,2 +1,0 @@
-# Copyright 2024 Bert24 authors
-# SPDX-License-Identifier: Apache-2.0
