+++ yamls/finetuning/glue/mosaic-bert-base-uncased.yaml
@@ -2,10 +2,10 @@
 # Please resolve any `null` fields before launching!
 
 # Whether to run the various GLUE jobs serially or in parallel (use parallel=True to take advantage of multiple GPUs)
-parallel: true
+parallel: false
 
 # Basic run configuration, additional details will be added to this name for each GLUE task, and each random seed
-base_run_name: mosaic-bert-base-uncased-glue-finetuning # Determines how runs are saved and logged in W&B
+base_run_name: cola-layernorm-test # Determines how runs are saved and logged in W&B
 default_seed: 19
 precision: amp_bf16
 
@@ -20,8 +20,8 @@ model:
 
 # Loading
 # (fill this in with the composer checkpoint from the end of pre-training a Mosaic BERT)
-starting_checkpoint_load_path: <checkpoint_from_pretraining>
-local_pretrain_checkpoint_folder: ./local-bert-checkpoints/
+# starting_checkpoint_load_path: <checkpoint_from_pretraining>
+# local_pretrain_checkpoint_folder: ./local-bert-checkpoints/
 
 # Saving
 save_finetune_checkpoint_prefix: ./local-finetune-checkpoints/ # (local)
@@ -29,9 +29,9 @@ save_finetune_checkpoint_prefix: ./local-finetune-checkpoints/ # (local)
 save_finetune_checkpoint_folder: ${save_finetune_checkpoint_prefix}/${base_run_name}
 
 # # (Optional) W&B logging
-# loggers:
-#   wandb:
-#     project:      # Fill this in if using W&B
+loggers:
+  wandb:
+    project: bert24      # Fill this in if using W&B
 #     entity:      # Fill this in if using W&B
 
 # Callbacks
@@ -45,40 +45,9 @@ scheduler:
   t_warmup: 0.06dur
   alpha_f: 0.0
 
-# Algorithms
-# algorithms:
-
 # Task configuration
 tasks:
-  mnli:
-    # Specify any extra task-specific arguments for the trainer here
-    trainer_kwargs:
-      # We keep one MNLI checkpoint locally so that we can start finetuning of
-      # RTE, MRPC and STS-B from the MNLI checkpoint
-      save_num_checkpoints_to_keep: 1
-  rte:
-    seeds: [19, 8364, 717, 10536, 90166]
-    trainer_kwargs:
-      save_num_checkpoints_to_keep: 0
-  qqp:
-    trainer_kwargs:
-      save_num_checkpoints_to_keep: 0
-  qnli:
-    trainer_kwargs:
-      save_num_checkpoints_to_keep: 0
-  sst2:
-    seeds: [19, 8364, 717]
-    trainer_kwargs:
-      save_num_checkpoints_to_keep: 0
-  stsb:
-    seeds: [19, 8364, 717, 10536, 90166]
-    trainer_kwargs:
-      save_num_checkpoints_to_keep: 0
-  mrpc:
-    seeds: [19, 8364, 717, 10536, 90166]
-    trainer_kwargs:
-      save_num_checkpoints_to_keep: 0
   cola:
-    seeds: [19, 8364, 717, 10536]
+    seeds: [19]
     trainer_kwargs:
       save_num_checkpoints_to_keep: 0
